Colored ouptut: error when `tput` binary is not available.
`tput` is used to check if the colored output is available or not. The `getPathToCommand` throw an `EnvironmentException` when a command can't be found.

This exception make the execution fail if the `tput` binary can't be found on the running machine. We can just return false if `tput` can't be found to disable colored output.